remove commented out functions from __init__.py
