+++ src/align_active_to_y_axis.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 
 class SNA_OT_Dgs_Render_Align_Active_To_Y_Axis_C305D(bpy.types.Operator):
     bl_idname = "sna.dgs_render_align_active_to_y_axis_c305d"
