+++ src/append_geometry_node_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Append_Geometry_Node_Modifier_C2492(bpy.types.Operator):
