+++ src/append_wire_cube.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 import os
 
 class SNA_OT_Dgs_Render_Append_Wire_Cube_56E0F(bpy.types.Operator):
