+++ src/append_wire_sphere.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 import os
