+++ src/apply_decimate_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 
 class SNA_OT_Dgs_Render_Apply_Decimate_Modifier_7A32C(bpy.types.Operator):
     bl_idname = "sna.dgs_render_apply_decimate_modifier_7a32c"
