+++ src/auto_generate_crop_object.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 import bpy.utils.previews
 import webbrowser
 import os
