+++ src/load_preview_icon.py
@@ -1,0 +1,12 @@
+import os
+
+_icons = None
+
+def load_preview_icon(path):
+    global _icons
+    if not path in _icons:
+        if os.path.exists(path):
+            _icons.load(path, path, "IMAGE")
+        else:
+            return 0
+    return _icons[path].icon_id
