+++ src/mesh23dgs_3dfed.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 import bpy.utils.previews
 import webbrowser
 import os
