+++ src/refresh_create_paint_attribute.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Refresh__Create_Paint_Attribute_84655(bpy.types.Operator):
