+++ src/remove_adjust_attribute_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Remove_Adjust_Attribute_Modifier_C5491(bpy.types.Operator):
