+++ src/remove_adjust_attributes_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Remove_Adjust_Attributes_Modifier_Fbc71(bpy.types.Operator):
