+++ src/remove_animate_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Remove_Animate_Modifier_5B34D(bpy.types.Operator):
