+++ src/remove_convert_to_rough_mesh_modifier.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 
 class SNA_OT_Dgs_Render_Remove_Convert_To_Rough_Mesh_Modifier_A8C4C(bpy.types.Operator):
