+++ src/remove_remove_by_size_modifier.py
@@ -1,10 +1,9 @@
 import bpy
 from .important import *
-from bpy_extras.io_utils import ImportHelper, ExportHelper
 
-class SNA_OT_Dgs_Render_Remove_Colour_Edit_Modifier_6255F(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_remove_colour_edit_modifier_6255f"
-    bl_label = "3DGS Render: Remove Colour Edit Modifier"
+class SNA_OT_Dgs_Render_Remove_Remove_By_Size_Modifier_3A0E5(bpy.types.Operator):
+    bl_idname = "sna.dgs_render_remove_remove_by_size_modifier_3a0e5"
+    bl_label = "3DGS Render: Remove Remove By Size Modifier"
     bl_description = "Removes the named modifier"
     bl_options = {"REGISTER", "UNDO"}
 
@@ -15,7 +14,7 @@ class SNA_OT_Dgs_Render_Remove_Colour_Edit_Modifier_6255F(bpy.types.Operator):
         return not False
 
     def execute(self, context):
-        bpy.context.view_layer.objects.active.modifiers.remove(modifier=bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Colour_Edit_GN'], )
+        bpy.context.view_layer.objects.active.modifiers.remove(modifier=bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Remove_By Size_GN'], )
         return {"FINISHED"}
 
     def invoke(self, context, event):
