+++ src/stop_interval_updates.py
@@ -1,4 +1,5 @@
 import bpy
+from .important import *
 
 class SNA_OT_Dgs_Render_Stop_Interval_Updates_5Ac80(bpy.types.Operator):
     bl_idname = "sna.dgs_render_stop_interval_updates_5ac80"
